+++ src/CleanRegex/Internal/Match/Details/DuplicateNamedGroupAdapter.php
@@ -3,7 +3,6 @@ namespace TRegx\CleanRegex\Internal\Match\Details;
 
 use TRegx\CleanRegex\Match\Details\Group\DuplicateNamedGroup;
 use TRegx\CleanRegex\Match\Details\Group\Group;
-use TRegx\CleanRegex\Match\Optional;
 
 class DuplicateNamedGroupAdapter implements DuplicateNamedGroup
 {
@@ -104,10 +103,5 @@ class DuplicateNamedGroupAdapter implements DuplicateNamedGroup
     public function or(string $substitute): string
     {
         return $this->group->or($substitute);
-    }
-
-    public function map(callable $mapper): Optional
-    {
-        return $this->group->map($mapper);
     }
 }