+++ src/CleanRegex/Replace/Details/Group/ReplaceGroup.php
@@ -2,7 +2,6 @@
 namespace TRegx\CleanRegex\Replace\Details\Group;
 
 use TRegx\CleanRegex\Match\Details\Group\Group;
-use TRegx\CleanRegex\Match\Optional;
 
 interface ReplaceGroup extends Group
 {
@@ -11,9 +10,4 @@ interface ReplaceGroup extends Group
     public function modifiedOffset(): int;
 
     public function byteModifiedOffset(): int;
-
-    /**
-     * @deprecated
-     */
-    public function map(callable $mapper): Optional;
 }