+++ test/Feature/CleanRegex/Match/Details/group/MatchDetailTest.php
@@ -3,14 +3,10 @@ namespace Test\Feature\TRegx\CleanRegex\Match\Details\group;
 
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
-use Test\Utils\ExampleException;
 use Test\Utils\ExplicitStringEncoding;
-use Test\Utils\Functions;
 use TRegx\CleanRegex\Exception\GroupNotMatchedException;
 use TRegx\CleanRegex\Exception\NonexistentGroupException;
 use TRegx\CleanRegex\Match\Details\Detail;
-use TRegx\CleanRegex\Match\Details\Group\Group;
-use TRegx\CleanRegex\Match\Details\Group\NotMatchedGroup;
 use function pattern;
 
 class MatchDetailTest extends TestCase
@@ -208,104 +204,6 @@ class MatchDetailTest extends TestCase
             ->match('Bar')
             ->first(function (Detail $detail) {
                 $detail->group('group')->text();
-            });
-    }
-
-    /**
-     * @test
-     */
-    public function shouldMapGroupOptional()
-    {
-        // when
-        $group = pattern('foo:(\w+)')->match('foo:bar')
-            ->first(function (Detail $detail) {
-                return $detail->group(1)
-                    ->map(function (Group $group) {
-                        return \strToUpper($group->text());
-                    })
-                    ->get();
-            });
-
-        // then
-        $this->assertSame('BAR', $group);
-    }
-
-    /**
-     * @test
-     */
-    public function shouldMapGroupOptionalEmptyFirst()
-    {
-        // then
-        $this->expectException(GroupNotMatchedException::class);
-        $this->expectExceptionMessage('Expected to get group #1, but the group was not matched');
-
-        // when
-        pattern('foo:(\w+)?')->match('foo:')
-            ->first(function (Detail $detail) {
-                return $detail->group(1)->map(Functions::fail())->get();
-            });
-    }
-
-    /**
-     * @test
-     */
-    public function shouldMapGroupOptionalEmptyFirstOrThrow()
-    {
-        // when
-        pattern('foo:(\w+)?')->match('foo:')->first(Functions::out($detail));
-        // then
-        $this->expectException(ExampleException::class);
-        $this->expectExceptionMessage('message');
-        // then
-        $detail->group(1)->map(Functions::fail())->orThrow(new ExampleException('message'));
-    }
-
-    /**
-     * @test
-     */
-    public function shouldMapGroupOptionalEmpty()
-    {
-        // then
-        $this->expectException(GroupNotMatchedException::class);
-        $this->expectExceptionMessage('Expected to get group #1, but the group was not matched');
-
-        // when
-        pattern('foo:(\w+)?')->match('foo:123 foo:')
-            ->forEach(function (Detail $detail) {
-                $detail->group(1)->map(Functions::identity())->get();
-            });
-    }
-
-    /**
-     * @test
-     */
-    public function shouldMapGroupOptionalUsingDuplicateName()
-    {
-        // given
-        pattern('foo:(?<name>\w+)')->match('foo:bar')->first(Functions::out($detail));
-        // when
-        $map = $detail->usingDuplicateName()->group('name')
-            ->map('\strToUpper');
-        $group = $map
-            ->get();
-        // then
-        $this->assertSame('BAR', $group);
-    }
-
-    /**
-     * @test
-     */
-    public function shouldOrElseRecieveNoArguments()
-    {
-        // when
-        pattern('\w+(?<group>Bar)?')
-            ->match('Lorem ipsum')
-            ->group('group')
-            ->stream()
-            ->forEach(function (NotMatchedGroup $group) {
-                $group
-                    ->map(Functions::identity())
-                    ->orElse(Functions::argumentless());
             });
     }
 }